Refactor prehog protos to match project conventions
