+++ build.assets/genproto.sh
@@ -10,7 +10,7 @@ main() {
   # Generated protos are written to
   # <teleport-root>/github.com/gravitational/teleport/..., so we copy them to
   # the correct relative path.
-  trap 'rm -fr github.com' EXIT   # don't leave github.com/ behind
+  trap 'rm -fr github.com' EXIT  # don't leave github.com/ behind
   rm -fr api/gen/proto gen/proto # cleanup gen/proto folders
 
   # Generate Gogo protos.
@@ -20,6 +20,10 @@ main() {
   # Generate protoc-gen-go protos.
   buf generate --template=buf-go.gen.yaml \
     --path=proto/teleport/lib/teleterm/
+
+  # Generate connect-go protos.
+  buf generate --template=buf-connect-go.gen.yaml \
+    --path=proto/prehog/
 
   # Generate JS protos.
   buf generate --template=buf-js.gen.yaml \
