--- lib/prehog/buf.gen.yaml
@@ -1,16 +1,0 @@
-version: v1
-managed:
-  enabled: true
-  go_package_prefix:
-    default: github.com/gravitational/teleport/lib/prehog/gen
-plugins:
-  - name: go
-    path: lib/prehog/bin/protoc-gen-go
-    out: lib/prehog/gen
-    opt:
-      - paths=source_relative
-  - name: connect-go
-    path: lib/prehog/bin/protoc-gen-connect-go
-    out: lib/prehog/gen
-    opt:
-      - paths=source_relative
