--- lib/prehog/proto/buf.yaml
@@ -1,6 +1,0 @@
-version: v1
-lint:
-  use:
-    - DEFAULT
-    - UNARY_RPC
-    - PACKAGE_NO_IMPORT_CYCLE
