+++ proto/prehog/v1alpha/teleport.proto
@@ -19,6 +19,8 @@ package prehog.v1alpha;
 import "google/protobuf/duration.proto";
 import "google/protobuf/timestamp.proto";
 
+option go_package = "github.com/gravitational/teleport/gen/proto/go/prehog/v1alpha";
+
 message UserLoginEvent {
   // anonymized
   string user_name = 1;
